+++ CHANGELOG.md
@@ -25,6 +25,7 @@ Usage:
 * Fix #1714: Add HelidonGenerator to add opinionated container image for Helidon applications
 * Fix #1985: Update outdated methods in Spring Boot CRD Maven Quickstart
 * Fix #2224: Quarkus native base image read from properties (configurable)
+* Fix #2228: Quarkus native base image uses UBI 8.7
 
 ### 1.13.1 (2023-06-16)
 * Fix #2212: Bump Kubernetes Client version to 6.7.2 (fixes issues when trace-logging OpenShift builds -regression in 6.7.1-)